+++ internal/pkg/codehub/repository.go
@@ -7,18 +7,18 @@ import (
 )
 
 type Repository interface {
-	AddStar(userID int, repoID int) error
-	DelStar(userID int, repoID int) error
-	GetStarredRepo(userID int) ([]gitmodels.Repository, error)
+	AddStar(userID int64, repoID int64) error
+	DelStar(userID int64, repoID int64) error
+	GetStarredRepo(userID int64) ([]gitmodels.Repository, error)
 
 	CreateIssue(issue models.Issue) error
 	UpdateIssue(issue models.Issue) error
-	CloseIssue(issueID int) error //закрывает вопрос(!не удаляет из бд)
-	GetIssuesList(repoID int) ([]models.Issue, error)
-	CheckAccessIssue(userID, issueID int) (perm.Permission, error) ///вернет "","read","update","close"
-	CheckAccessRepo(userID, repoID int) (perm.Permission, error)
+	CloseIssue(issueID int64) error //закрывает вопрос(!не удаляет из бд)
+	GetIssuesList(repoID int64) ([]models.Issue, error)
+	CheckAccessIssue(userID, issueID int64) (perm.Permission, error) ///вернет "","read","update","close"
+	CheckAccessRepo(userID, repoID int64) (perm.Permission, error)
 	//обозначение по порядку(не может читать, может
 	//просмотреть, может изменить(коллаборатор), может
 	//закрыть(автор или админ репозитория)
-	GetIssues(issueID int) (models.Issue, error)
+	GetIssues(issueID int64) (models.Issue, error)
 }
