+++ internal/pkg/codehub/usecase.go
@@ -6,10 +6,10 @@ import (
 
 type UCCodeHub interface {
 	ModifyStar(star models.Star) error
-	GetStarredRepo(userID int) (models.RepoSet, error)
+	GetStarredRepo(userID int64) (models.RepoSet, error)
 	CreateIssue(issue models.Issue) error
 	UpdateIssue(issue models.Issue) error
-	CloseIssue(issueID int, userID int) error
-	GetIssuesList(repoID int, userID int) (models.IssuesSet, error)
-	GetIssue(issueID int, userID int) (models.Issue, error)
+	CloseIssue(issueID int64, userID int64) error
+	GetIssuesList(repoID int64, userID int64) (models.IssuesSet, error)
+	GetIssue(issueID int64, userID int64) (models.Issue, error)
 }
