+++ internal/pkg/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       int    `json:"id" valid:"-" db:"id"`
+	ID       int64    `json:"id" valid:"-" db:"id"`
 	Password string `json:"password,omitempty" valid:"stringlength(5|128)" db:"password"`
 	Name     string `json:"name" valid:"optional,alphanum" db:"name"`
 	Login    string `json:"login" valid:"stringlength(3|128)" db:"login"`
@@ -13,3 +13,4 @@ type SignInForm struct {
 	Login    string `json:"login" valid:"alphanum,stringlength(3|128)"`
 	Password string `json:"password" valid:"stringlength(5|128)"`
 }
+//easyjson -all path/to/file.go
